requirements.txt
numpy==1.23.5
Pillow==9.3.0
matplotlib==3.6.2
tk==0.1.0
